fix: Fix timestamp unit (#226)
Keep the block header timestamp in seconds, while using microseconds for ordered blocks to pass to metadata transactions.

+++ crates/pipe-exec-layer-ext-v2/execute/src/onchain_config/types.rs
@@ -66,13 +66,13 @@ sol! {
         function blockPrologue(
             bytes proposer,
             uint64[] calldata failedProposerIndices,
-            uint256 timestampMicros
+            uint64 timestampMicros
         );
 
         function blockPrologueExt(
             bytes proposer,
             uint64[] calldata failedProposerIndices,
-            uint256 timestampMicros
+            uint64 timestampMicros
         );
 }
